#726: Extract edit-side functions into helper
